+++ apps/web/src/app/recruit/[id]/page.tsx
@@ -1,8 +1,8 @@
 import Link from 'next/link';
 import { FC } from 'react';
 
-import { BlockRenderer } from '@/components/notion/renderer';
-import { recruitBlockComponents } from '@/components/recruit/recruitBlockComponents';
+import BlockRenderer from '@/components/notion/unofficial/BlockRenderer';
+import { recruitBlockComponents } from '@/components/recruit/recruitBlockUnofficial';
 import { gateway } from '@/gateway';
 
 interface RecruitSubPageProps {
@@ -12,13 +12,23 @@ interface RecruitSubPageProps {
 const RecruitSubPage: FC<RecruitSubPageProps> = async ({ params }) => {
   const pageId = params.id;
 
-  const { blocks } = await gateway.recruit.page.query({ id: pageId });
+  const { title, pageBlock, blockMap } = await gateway.recruit.page.query({ id: pageId });
+
+  function getBlock(id: string) {
+    const block = blockMap[id];
+    if (block) {
+      return block;
+    }
+    throw new Error('Invalid Block Id: ' + id);
+  }
 
   return (
     <div>
+      {title}
       <BlockRenderer
-        blocks={blocks}
+        blocks={pageBlock.content ?? []}
         blockComponents={recruitBlockComponents}
+        getBlock={getBlock}
         renderPageLink={({ id, name, className }) => (
           <Link href={`/recruit/${id}`} className={className}>
             {name}
